specify timings as floats, so more fine grained adjustments are possible.
Wed Feb  6 04:54:16 2002  Tim Janik  <timj@gtk.org>

        * bsesequencer.c: specify timings as floats, so more
        fine grained adjustments are possible.

svn path=/trunk/; revision=359

+++ plugins/bsesequencer.h
@@ -46,7 +46,7 @@ struct _BseSequencer
 {
   BseSource      parent_object;
 
-  guint		 counter;
+  gfloat	 counter;
   gint		 transpose;
   BseSequence	*sdata;
   guint		 n_freq_values;
